+++ udata_ldap/commands.py
@@ -7,6 +7,7 @@ from click import echo, prompt
 from flask_ldap3_login import AuthenticationResponseStatus
 
 from udata.commands import cli, success, white, exit_with_error, header
+from udata.utils import safe_unicode
 
 from .ldap import manager
 
@@ -27,7 +28,7 @@ def config():
     for key in sorted(manager.config):
         if key.startswith('LDAP_'):
             echo('{key}: {value}'.format(key=white(key),
-                                         value=manager.config[key]))
+                                         value=safe_unicode(manager.config[key])))
 
 
 @ldap.command()
@@ -40,7 +41,7 @@ def check():
     try:
         who_am_i = manager.connection.extend.standard.who_am_i()
         success('Bind DN successfully connected')
-        echo('Bind DN user is "{}"'.format(white(who_am_i)))
+        echo('Bind DN user is "{}"'.format(white(safe_unicode(who_am_i))))
     except Exception as e:
         exit_with_error('Unable to connect', e)
 
@@ -51,17 +52,17 @@ def check():
     if result.status == AuthenticationResponseStatus.success:
         success('User successfully connected')
         echo('Authenticated user is "{email} ({dn})"'.format(
-            email=white(result.user_id),
-            dn=white(result.user_dn)
+            email=white(safe_unicode(result.user_id)),
+            dn=white(safe_unicode(result.user_dn))
         ))
         echo('User has the following remote attributes:')
         for key, value in result.user_info.items():
             echo('{key}: {value}'.format(key=white(key),
-                                         value=value))
+                                         value=safe_unicode(value)))
         echo('Local user will be createdwith the following values:')
         for key, value in manager.extract_user_infos(result.user_info).items():
             echo('{key}: {value}'.format(key=white(key),
-                                         value=value))
+                                         value=safe_unicode(value)))
     else:
         exit_with_error('Unable to authenticate user "{0}"'.format(email))