+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Display errors on login form
 - Force email into the login form
 - Fix encoding errors in ldap commands
+- Update user on login
+- Start handling errors on negociate view
 
 ## 0.2.1 (2018-10-08)