+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Update user on login
 - Start handling errors on negociate view
 - Display a page when trying automatic login wihtout credentials
+- Adds translations
 
 ## 0.2.1 (2018-10-08)